+++ apps/ubike/src/services/local.ts
@@ -1,10 +1,17 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
 import { SuccessApiResponse } from 'next-api-handler';
 
+import type {
+  BikeCyclingWithGeoJson,
+  CyclingQueryParam,
+} from '@/pages/api/cyclings';
 import type { StationQueryParam, StationWithBike } from '@/pages/api/stations';
+
+const ONE_HOUR = 60 * 60;
 
 export const localApi = createApi({
   reducerPath: 'local',
+  keepUnusedDataFor: ONE_HOUR,
   baseQuery: fetchBaseQuery({
     baseUrl: '/api',
   }),
@@ -15,7 +22,13 @@ export const localApi = createApi({
     >({
       query: ({ lat, lng }) => `stations?lat=${lat}&lng=${lng}`,
     }),
+    getCyclingByCity: builder.query<
+      SuccessApiResponse<BikeCyclingWithGeoJson[]>,
+      CyclingQueryParam['city']
+    >({
+      query: (citySlug) => `cyclings?city=${citySlug}`,
+    }),
   }),
 });
 
-export const { useLazyGetStationsQuery } = localApi;
+export const { useLazyGetStationsQuery, useGetCyclingByCityQuery } = localApi;