+++ libs/ptx/src/bikes.ts
@@ -1,4 +1,5 @@
 import { apiGet, ApiParam } from './lib/api';
+import { City, PTXCityMap } from './lib';
 
 export interface Station {
   StationUID: string;
@@ -47,6 +48,21 @@ export interface BikeQueryParam {
   count: number;
 }
 
+export interface BikeCycling {
+  RouteName: string;
+  AuthorityName?: string;
+  CityCode: string;
+  City: City;
+  Town?: string;
+  RoadSectionStart?: string;
+  RoadSectionEnd?: string;
+  Direction?: string;
+  CyclingLength?: number;
+  UpdateTime: string;
+  Geometry: string;
+  FinishedTime?: string;
+}
+
 const mapToPTXParam = ({
   lat,
   lng,
@@ -63,3 +79,9 @@ export const getNearByStations = (query: BikeQueryParam) =>
 
 export const getNearByAvailableBikes = (query: BikeQueryParam) =>
   apiGet<AvailableBike[]>('Bike/Availability/NearBy', mapToPTXParam(query));
+
+export const getCyclingShapeByCity = (city: City) =>
+  apiGet<BikeCycling[]>(`Cycling/Shape/${PTXCityMap[city]}`, {
+    $top: '100',
+    $filter: 'CyclingLength gt 0',
+  });